+++ backend/src/emitters/index.js
@@ -1,0 +1,5 @@
+const user = require("./users");
+const game = require("./game");
+const messages = require("./messages");
+
+module.exports = { user, game, messages };
