+++ backend/src/emitters/messages.js
@@ -1,0 +1,5 @@
+exports.chatMessage = (io, gameId, message, sender) => {
+  io.of("/game")
+    .in(`${gameId}`)
+    .emit("chat message", { message, sender: sender.nickname });
+};
