+++ backend/src/events/index.js
@@ -1,0 +1,5 @@
+const game = require("./game");
+const users = require("./users");
+const messages = require("./messages");
+
+module.exports = { game, users, messages };
